+++ src/api/src/main/kotlin/com/tenmm/tilserver/post/adapter/inbound/rest/model/DeletePostResponse.kt
@@ -1,4 +1,4 @@
-package com.tenmm.tilserver.post.adapter.inbound.model
+package com.tenmm.tilserver.post.adapter.inbound.rest.model
 
 data class DeletePostResponse(
     val isSuccess: Boolean,
