+++ src/api/src/main/kotlin/com/tenmm/tilserver/post/adapter/inbound/rest/model/GetPostListResponse.kt
@@ -1,4 +1,4 @@
-package com.tenmm.tilserver.post.adapter.inbound.model
+package com.tenmm.tilserver.post.adapter.inbound.rest.model
 
 import com.tenmm.tilserver.post.application.inbound.model.GetPostListResult
 import com.tenmm.tilserver.post.domain.Post
